+++ internal/apiclient/model_list_patch_body.go
@@ -22,6 +22,8 @@ var _ MappedNullable = &ListPatchBody{}
 type ListPatchBody struct {
 	// The new name of the list.
 	Name *string `json:"name,omitempty"`
+	// The description of the list.
+	Description NullableString `json:"description,omitempty"`
 	// The items in the list.
 	Add []ListItemsInner `json:"add,omitempty"`
 	// Items to remove from the list.
@@ -80,6 +82,49 @@ func (o *ListPatchBody) HasName() bool {
 // SetName gets a reference to the given string and assigns it to the Name field.
 func (o *ListPatchBody) SetName(v string) {
 	o.Name = &v
+}
+
+// GetDescription returns the Description field value if set, zero value otherwise (both if not set or set to explicit null).
+func (o *ListPatchBody) GetDescription() string {
+	if o == nil || IsNil(o.Description.Get()) {
+		var ret string
+		return ret
+	}
+	return *o.Description.Get()
+}
+
+// GetDescriptionOk returns a tuple with the Description field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+// NOTE: If the value is an explicit nil, `nil, true` will be returned
+func (o *ListPatchBody) GetDescriptionOk() (*string, bool) {
+	if o == nil {
+		return nil, false
+	}
+	return o.Description.Get(), o.Description.IsSet()
+}
+
+// HasDescription returns a boolean if a field has been set.
+func (o *ListPatchBody) HasDescription() bool {
+	if o != nil && o.Description.IsSet() {
+		return true
+	}
+
+	return false
+}
+
+// SetDescription gets a reference to the given NullableString and assigns it to the Description field.
+func (o *ListPatchBody) SetDescription(v string) {
+	o.Description.Set(&v)
+}
+
+// SetDescriptionNil sets the value for Description to be an explicit nil
+func (o *ListPatchBody) SetDescriptionNil() {
+	o.Description.Set(nil)
+}
+
+// UnsetDescription ensures that no value is present for Description, not even an explicit nil
+func (o *ListPatchBody) UnsetDescription() {
+	o.Description.Unset()
 }
 
 // GetAdd returns the Add field value if set, zero value otherwise.
@@ -191,6 +236,9 @@ func (o ListPatchBody) ToMap() (map[string]interface{}, error) {
 	if !IsNil(o.Name) {
 		toSerialize["name"] = o.Name
 	}
+	if o.Description.IsSet() {
+		toSerialize["description"] = o.Description.Get()
+	}
 	if !IsNil(o.Add) {
 		toSerialize["add"] = o.Add
 	}
@@ -223,6 +271,7 @@ func (o *ListPatchBody) UnmarshalJSON(data []byte) (err error) {
 
 	if err = json.Unmarshal(data, &additionalProperties); err == nil {
 		delete(additionalProperties, "name")
+		delete(additionalProperties, "description")
 		delete(additionalProperties, "add")
 		delete(additionalProperties, "remove")
 		delete(additionalProperties, "labels")