+++ internal/apiclient/model_lists_items_inner.go
@@ -26,6 +26,8 @@ type ListsItemsInner struct {
 	Id string `json:"id"`
 	// The name of the list.
 	Name string `json:"name"`
+	// The description of the list.
+	Description *string `json:"description,omitempty"`
 	// User ID of the user that created the list.
 	CreatedBy string `json:"created_by"`
 	// The date the list was created.
@@ -113,6 +115,38 @@ func (o *ListsItemsInner) GetNameOk() (*string, bool) {
 // SetName sets field value
 func (o *ListsItemsInner) SetName(v string) {
 	o.Name = v
+}
+
+// GetDescription returns the Description field value if set, zero value otherwise.
+func (o *ListsItemsInner) GetDescription() string {
+	if o == nil || IsNil(o.Description) {
+		var ret string
+		return ret
+	}
+	return *o.Description
+}
+
+// GetDescriptionOk returns a tuple with the Description field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *ListsItemsInner) GetDescriptionOk() (*string, bool) {
+	if o == nil || IsNil(o.Description) {
+		return nil, false
+	}
+	return o.Description, true
+}
+
+// HasDescription returns a boolean if a field has been set.
+func (o *ListsItemsInner) HasDescription() bool {
+	if o != nil && !IsNil(o.Description) {
+		return true
+	}
+
+	return false
+}
+
+// SetDescription gets a reference to the given string and assigns it to the Description field.
+func (o *ListsItemsInner) SetDescription(v string) {
+	o.Description = &v
 }
 
 // GetCreatedBy returns the CreatedBy field value
@@ -287,6 +321,9 @@ func (o ListsItemsInner) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
 	toSerialize["id"] = o.Id
 	toSerialize["name"] = o.Name
+	if !IsNil(o.Description) {
+		toSerialize["description"] = o.Description
+	}
 	toSerialize["created_by"] = o.CreatedBy
 	toSerialize["created_at"] = o.CreatedAt
 	toSerialize["kind"] = o.Kind
@@ -347,6 +384,7 @@ func (o *ListsItemsInner) UnmarshalJSON(data []byte) (err error) {
 	if err = json.Unmarshal(data, &additionalProperties); err == nil {
 		delete(additionalProperties, "id")
 		delete(additionalProperties, "name")
+		delete(additionalProperties, "description")
 		delete(additionalProperties, "created_by")
 		delete(additionalProperties, "created_at")
 		delete(additionalProperties, "kind")